Nwt 11 notifications

+++ app/src/main/java/com/meadowlandapps/simplelists/model/CategoryModel.kt
@@ -1,5 +1,5 @@
 package com.meadowlandapps.simplelists.model
 
-class CategoryModel(var id: Int,
+class CategoryModel(var id: Long,
                     var name: String,
                     var selected: Boolean)
